% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_sp500_data}
\alias{get_sp500_data}
\title{Downloads stock data from yahoo finance}
\usage{
get_sp500_data(
  stock_names = sp500_names,
  from = "2019-11-27",
  to = Sys.Date(),
  remove_missing = TRUE,
  show_output = FALSE,
  bind = FALSE
)
}
\arguments{
\item{stock_names}{Vector including the symbols of stocks we want to retrieve}

\item{from}{Date to get data after}

\item{to}{Date to get data until}

\item{remove_missing}{Removes the data with missing rows}
}
\value{
List of dataframes containg data for each sp500 company
}
\description{
Downloads stock data from yahoo finance
}
