% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{load_sp500_data}
\alias{load_sp500_data}
\title{Loads data from a csv file}
\usage{
load_sp500_data(dir_path)
}
\arguments{
\item{dir_path}{Directory to load data from}
}
\value{
List of dataframes for each file contained in dir_path
}
\description{
Loads data from a csv file
}
