% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_class.R
\docType{class}
\name{portfolio-class}
\alias{portfolio-class}
\alias{portfolio}
\title{Reference class for portfolio object}
\description{
Reference class for portfolio object
}
\section{Fields}{

\describe{
\item{\code{assets}}{n-list of asset symbols e.g. c("AAPL", "KO")}

\item{\code{data}}{A single data frame containing each asset}

\item{\code{shares}}{n-vector of numbers of shares for each asset}

\item{\code{holdings}}{n-vector of holdings for each asset}

\item{\code{time}}{Current time of the portfolio (time points from start)}

\item{\code{value}}{Total value of the portfolio}

\item{\code{weights}}{n-vector of asset weightings}

\item{\code{T}}{Number of data points for each asset}

\item{\code{investment}}{Amount invested in the portfolio}

\item{\code{returns_matrix}}{Matrix of assets by asset percentage returns}

\item{\code{asset_summary}}{Summary statistics for assets}
}}


