% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{save_sp500_data}
\alias{save_sp500_data}
\title{Saves a list of data frames to csv files}
\usage{
save_sp500_data(data, dir_path)
}
\arguments{
\item{data}{List of dataframes to save as csv files}

\item{dir_path}{Directory that the csv files will be written to}
}
\description{
Saves a list of data frames to csv files
}
